requirements.txt
pystray
Pillow
keyboard
pywin32
winotify
